Add `buildkite-agent` rootfs image
This contains the rootfs that we use to run the actual `buildkite-agent`
process within.  It's much more minimal than most of these rootfs
images, and is purposefully so, to encourage CI passes to be run within
a pass-specific rootfs image.

This also beefs up the rootfs construction utilities somewhat,
incorporating some of the best practices scattered around various
similar efforts.